DeepBanana.Exception

Description

A stab at an extensible exceptions system, used internally. To be displaced eventually by whatever "composable effects" library becomes the de facto standard in the Haskell community.

Extensible ariants of exception datatypes

data Coproduct l where

Coproduct of datatypes, indexed by a type-level lists of these datatypes. For instance, Coproduct '[a,b] is equivalent to Either a b.

Constructors

Left' :: a -> Coproduct (a : l)

Right' :: Coproduct l -> Coproduct (a : l)

class Variant t e where

Type class to deal with variant datatypes in a generic fashion.

Methods

setVariant :: e -> t

getVariant :: t -> Maybe e

throwVariant :: (MonadError t m, Variant t e) => e -> m a

Throwing an exception when the error datatype is a variant including that datatype.

catchVariant :: forall m t e a. (MonadError t m, Variant t e) => m a -> (e -> m a) -> m a

Catching one possible exception type out of a variant.

Stack traces

data WithStack e

Simple wrapper adding a call stack to an arbitrary exception.

withStack :: (?loc :: CallStack) => e -> WithStack e

Managing exception monads

unsafeRunExcept :: Exception e => Either e a -> a

Gets the Right part of a computation, throwing an imprecise exception in case of Left.

embedExcept :: MonadError e m => Either e a -> m a

Embeds an Either value into an arbitrary error monad.

runExceptTAs :: Proxy t -> ExceptT t m a -> m (Either t a)

Synonym for ExceptT specifying the exception type via a Proxy. Useful to disambiguate types for the compiler.

Specific handlers

attemptGCThenRetryOn :: forall m t e a. (MonadIO m, MonadError t m, Variant t e) => Proxy e -> m a -> m a

Runs a computation, and in case of a specific exception being raised performs garbage collection before attempting the computation again. This is useful when allocating a lot of memory quickly on the GPU, as the garbage collector may not have enough time to run by itself. Throws back the same exception if the second run fails anyway. This is used anytime an allocation is performed on a GPU device, most notably emptyTensor and other tensor creation functions.