DeepBanana.HList

Description

Heterogeneous lists. Used notably for:

Tensor shapes

Neural network weights

Multi GPU

It is similar in principle and very much inspired by Oleg Kiselyov's HList package.

Heterogeneous list datatypes

data family HList l

An HList l is an heterogeneous list indexed by a type-level list l of its elements datatypes.

1:."Hi!":.(+):.Z :: HList '[Int,String,Float -> Float -> Float]

type SizedList n a = HList (SizedList' n a)

Fixed size homogeneous list datatype. Useful to define shapes with dynamic dimensions and fixed length, for instance.

type family SizedList' n a

Helper type family for SizedList.

Equations

SizedList' 0 a = `[]`

SizedList' n a = a : SizedList' (n - 1) a

Concatenating heterogeneous lists

class Concat l1 l2 where

Concatenating and splitting heterogeneous lists.

Associated Types

type ConcatRes l1 l2 :: [*]

Methods

hconcat :: HList l1 -> HList l2 -> HList (ConcatRes l1 l2)

hsplit :: HList (ConcatRes l1 l2) -> (HList l1, HList l2)

Heterogeneous categories and monoids

class HCategory cat where

Heterogeneous category type class. Used to overload feed-forward composition across a few layer datatypes.

Methods

id' :: cat `[]` a a

(>+>) :: Concat l1 l2 => cat l1 a b -> cat l2 b c -> cat (ConcatRes l1 l2) a c infixr 3

class HMonoid t where

Heterogeneous monoid type class. Used to overload heterogeneous list concatenation for weights datatypes notably.

Methods

hmempty :: t `[]`

(<+>) :: Concat l1 l2 => t l1 -> t l2 -> t (ConcatRes l1 l2) infixr 3