DeepBanana.Layer

Description

Compositional layers which learn a set of weights through backpropagation.

Documentation

module DeepBanana.Weights

newtype Layer m a w inp out

Constructors

Layer

Fields

forwardBackward :: Weights a w -> inp -> m (out, out -> (Weights a w, inp))

data InitLayer m a w inp out

initWeights :: InitLayer m a w inp out -> m (Weights a w)

layer :: InitLayer m a w inp out -> Layer m a w inp out

init :: Layer m a w inp out -> m (Weights a w) -> InitLayer m a w inp out

init' :: Applicative m => Layer m a `[]` inp out -> InitLayer m a `[]` inp out

forward :: Monad m => Layer m a w inp out -> Weights a w -> inp -> m out

Runs the forward pass of a layer.

backward :: Monad m => Layer m a w inp out -> Weights a w -> inp -> out -> m (Weights a w, inp)

Runs the backward pass of a layer.

combinePasses :: Monad m => (Weights a w -> inp -> m out) -> (Weights a w -> inp -> out -> m (out -> (Weights a w, inp))) -> Layer m a w inp out

Creates a layer from separate forward pass and backward pass computations.

combinePasses' :: Monad m => (inp -> m out) -> (inp -> out -> m (out -> inp)) -> Layer m a `[]` inp out

(***) :: forall m s w1 w2 a b a' b' n. (Monad m, Concat w1 w2) => Layer m s w1 a b -> Layer m s w2 a' b' -> Layer m s (ConcatRes w1 w2) (a, a') (b, b') infixr 4

(&&&) :: forall m s w1 w2 a b b' n. (Monad m, Concat w1 w2, AdditiveGroup a) => Layer m s w1 a b -> Layer m s w2 a b' -> Layer m s (ConcatRes w1 w2) a (b, b') infixr 4

first :: (AdditiveGroup b, Monad m) => Layer m s `[]` (a, b) a

second :: (AdditiveGroup a, Monad m) => Layer m s `[]` (a, b) b

terminal :: (AdditiveGroup inp, Monad m) => out -> Layer m a `[]` inp out

(-<) :: forall m i1 i2 w a out. (Monad m, VectorSpace i1, VectorSpace i2) => Layer m a w (i1, i2) out -> i1 -> Layer m a w i2 out infixr 4

noWeights :: Monad m => (inp -> m (out, out -> inp)) -> Layer m a `[]` inp out

effect :: Monad m => (a -> m b) -> Layer m s `[]` a a