DeepBanana.Tensor

Description

GPU multi-dimensional dense numeric arrays.

Documentation

module DeepBanana.Tensor.Shape

module DeepBanana.Tensor.TensorScalar

module Data.VectorSpace

data Tensor s a

An immutable, GPU tensor with a given shape s, whose scalar type a should be an instance of TensorScalar. Such tensors are instances of Num, Fractional and Floating, implementing all the numeric operations in an elementwise fashion:

>>> fromList [1,2,3] + fromList [3,4,5] :: Tensor '[3] CFloat
Tensor [3] [4.0,6.0,8.0]

Scalars are automatically broadcast to whatever shape is necessary:

>>> 5 * fromList [1,2,3] :: Tensor '[3] CFloat
Tensor [3] [5.0,10.0,15.0]

Constructors

Tensor (ForeignPtr a)

Basic manipulations

dtype :: forall s a. TensorScalar a => Tensor s a -> DataType

Returns the CuDNN datatype of a tensor.

reshape :: (Shape s1, Shape s2, Size s1 ~ Size s2) => Tensor s1 a -> Tensor s2 a

Type safe reshaping.

Converting from/to mutable tensors

unsafeFreeze :: PrimMonad m => MTensor (PrimState m) s a -> m (Tensor s a)

Converts a mutable tensor into an immutable one in O(1) time. The data is not copied, so one should make sure the input mutable tensor is not modified after the the conversion. Unsafe.

unsafeThaw :: PrimMonad m => Tensor s a -> m (MTensor (PrimState m) s a)

Converts an immutable tensor into a mutable one in O(1) time. The data is not copied, so one should make sure the input immutable tensor is not used after the conversion. Unsafe.

Converting from/to lists

fromList :: (TensorScalar a, Shape s) => [a] -> Tensor s a

Initializes a tensor from list data. If the length of the list does not correspond to the size of the desired shape, throws an exception at run time.

toList :: (TensorScalar a, Shape s) => Tensor s a -> [a]

Converts a tensor to a list.

Converting from/to storable vectors

fromVector :: forall s a. (TensorScalar a, Shape s) => Vector a -> Tensor s a

Converts a storable vector to an immutable tensor. Throws an error at runtime when the input's length does not correspond to the desired output size. Runs in O(n) time.

toVector :: forall a s. (TensorScalar a, Shape s) => Tensor s a -> Vector a

Converts a tensor to a storable vector. Runs in O(n) time.

Utilities

tconcat :: forall n m a. (KnownNat n, KnownNat m, KnownNat (n + m), TensorScalar a) => Tensor `[n]` a -> Tensor `[m]` a -> Tensor `[n + m]` a

Concatenates two 1-dimensional tensors. Inverse of tsplit. Runs in O(n+m) time.

tsplit :: forall n m a. (KnownNat n, KnownNat m, TensorScalar a) => Tensor `[n + m]` a -> (Tensor `[n]` a, Tensor `[m]` a)

Splits a 1-dimensional tensor into 2 parts. Inverse of tconcat. Runs in O(n+m) time.

elementwiseMax :: forall s a. (Shape s, TensorScalar a) => Tensor s a -> Tensor s a -> Tensor s a

Computes the elementwise maximum of 2 tensors.