DeepBanana.Tensor.Mutable

Description

Mutable GPU multi-dimensional dense numeric arrays, parametrized by the state token of some primitive state monad.

Documentation

data MTensor st d s a

Mutable tensor, parametrized by some state token st, stored on a device d, with shape s and scalar datatype a.

Constructors

MTensor

Fields

shape :: s

dataptr :: ForeignPtr a

type IOTensor = MTensor RealWorld

Convenient synonym for mutable tensors in the IO monad.

dtype :: forall m d n a. (PrimMonad m, TensorScalar a) => MTensor (PrimState m) d n a -> m DataType

Returns the CuDNN datatype identifier for the tensor.

Creating mutable tensors

emptyTensor :: forall s d m t a. (Shape s, Device d, TensorScalar a, PrimMonad m, MonadError t m, Variant t OutOfMemory) => s -> m (MTensor (PrimState m) d s a)

Creates an uninitialized mutable tensor with a given shape. Uninitialized means the contents of the tensor are defined by whatever was lying around in memory at the time. In case there is not enough space on the device, will throw and OutOfMemory exception after attempting garbage collection first.

tensorFromList :: forall m d s a t. (PrimMonad m, Device d, Shape s, TensorScalar a, MonadError t m, Variant t OutOfMemory, Variant t IncompatibleSize) => s -> [a] -> m (MTensor (PrimState m) d s a)

Creates a mutable tensor from a list. Throws an exception at runtime when the size of the list does not correspond to the desired shape.

zeros :: (PrimMonad m, Device d, Shape s, TensorScalar a, MonadError t m, Variant t OutOfMemory) => s -> m (MTensor (PrimState m) d s a)

Creates a mutable tensor filled with zeros.

ones :: (PrimMonad m, Device d, Shape s, TensorScalar a, MonadError t m, Variant t OutOfMemory) => s -> m (MTensor (PrimState m) d s a)

Creates a mutable tensor filled with ones.

Basic manipulations

withDevicePtr :: Storable a => IOTensor d s a -> (DevicePtr a -> IO b) -> IO b

Runs an IO action with the internal device pointer of a given mutable tensor.

reshape :: (Shape s1, Shape s2, MonadError t m, Variant t IncompatibleSize) => s2 -> MTensor st d s1 a -> m (MTensor st d s2 a)

Tensor reshaping. Will throw and IncompatibleSize exception if the desired shape size does not correspond to the input tensor size.

reshape' :: forall d s1 s2 st a. (Shape s1, Shape s2) => s2 -> MTensor st d s1 a -> MTensor st d s2 a

Unsafe pure tensor reshaping, which throws an imprecise exception when the desired shape size does not correspond to the input tensor size.

copy :: forall m d1 d2 s a t. (PrimMonad m, Device d2, Shape s, TensorScalar a, MonadError t m, Variant t OutOfMemory) => MTensor (PrimState m) d1 s a -> m (MTensor (PrimState m) d2 s a)

Copies the data of a mutable tensor into a new one.

tensorToList :: forall m d s a. (PrimMonad m, Device d, Shape s, TensorScalar a) => MTensor (PrimState m) d s a -> m [a]

Converts a mutable tensor to a list of elements.

In place numeric functions

threshInplace :: forall m d s a. (PrimMonad m, Device d, Shape s, TensorScalar a) => MTensor (PrimState m) d s a -> a -> m ()

Thresholds a mutable tensor in place. Used to implement dropout.

tlog :: forall m d s a. (PrimMonad m, Device d, Shape s, TensorScalar a) => MTensor (PrimState m) d s a -> m ()

In place logarithm.

texp :: forall m d s a. (PrimMonad m, Device d, Shape s, TensorScalar a) => MTensor (PrimState m) d s a -> m ()

In place exponential.

inv :: forall m d s a. (PrimMonad m, Device d, Shape s, TensorScalar a) => MTensor (PrimState m) d s a -> m ()

In place inverse.

Re-exports

module DeepBanana.Tensor.Shape

module DeepBanana.Tensor.TensorScalar