DeepBanana.Tensor.Shape

Description

Shape datatypes for tensors.

Shape datatypes

type Dim n = SizedList n Int

Dim n, where n is a type level natural, is a dynamic n-dimensional shape. It is represented by an HList of fixed size n where all elements are integers.

type Fixed l = HList (Fixed' l)

Fixed l, where l is a type level list of type level natural numbers, represents a shape which is fully fixed at the type level. It is represented by an HList where all elements are Proxy n for some type level natural n.

Shape type classes

class (Eq s, Ord s, Show s, Serialize s, NFData s) => Shape s where

Type class that all shape datatypes implement.

Minimal complete definition

dimensions, scalarShape

Methods

dimensions :: s -> [Int]

Dimensions of the shape as a list of ints.

scalarShape :: s

Smallest shape of the datatype, meaning it can be broadcasted against any other shape of the datatype.

nbdim :: s -> Int

Number of dimensions.

size :: s -> Int

Size, i.e. the product of the dimensions.

broadcastable :: (Shape s1, Shape s2) => s1 -> s2 -> Bool

Returns true if and only if the first shape argument can be broadcasted to to the second shape argument. Currently, broadcasting is limited to shapes holding the same number of dimensions. Then, a dimension d1 is broadcastable to a dimension d2 if an only if d1 == d2 or d1 == 1.

class FixedShape l where

Utility type class for shapes which are fully defined at the type level.

Associated Types

type Fixed' l :: [*]

type Nbdim l :: Nat

type Size l :: Nat

Methods

fixed :: Proxy l -> HList (Fixed' l)

Utility function to define shapes fixed at the type level.

class Shape s => IsFixed s

Utility type class for shapes which are fully fixed at the type level.

Converting from fixed to dynamic shape

data AnyFixed where

Existential datatype wrapping any fixed shape. Useful when converting from a fixed type level shape to a dynamic shape locally.

Constructors

AnyFixed :: IsFixed (HList l) => HList l -> AnyFixed

toAnyFixed :: Shape s => s -> AnyFixed

Converts an arbitrary shape to its fixed version.

Re-exports

module DeepBanana.HList