DeepBanana.Tensor.TensorScalar

Description

Defines datatype which can be used as scalar elements for a tensor. Also exports much needed, although orphan instances for CFloat and CDouble: Generic, NFData and Serialize.

Documentation

class (Floating a, Storable a, VectorSpace a, a ~ Scalar a, Serialize a, NFData a) => TensorScalar a where

Type class for tensor scalars. Basically requires them to be storable, serializable, deepseqable numbers that can be handled by CUDA, Cublas, CuDNN and Haskell. Basically requires low-level numeric routines.

Methods

datatype :: Proxy a -> DataType

The corresponding CuDNN datatype.

thresh :: DevicePtr a -> CSize -> a -> DevicePtr a -> DeviceM d ()

Low-level in-place numeric operations.

rawMul :: DevicePtr a -> DevicePtr a -> CSize -> DeviceM d ()

rawAdd :: DevicePtr a -> DevicePtr a -> CSize -> DeviceM d ()

rawAbs :: DevicePtr a -> CSize -> DeviceM d ()

rawSignum :: DevicePtr a -> CSize -> DeviceM d ()

rawSubtract :: DevicePtr a -> DevicePtr a -> CSize -> DeviceM d ()

rawNegate :: DevicePtr a -> CSize -> DeviceM d ()

rawScale :: a -> DevicePtr a -> CSize -> DeviceM d ()

rawLog :: DevicePtr a -> CSize -> DeviceM d ()

rawInv :: DevicePtr a -> CSize -> DeviceM d ()

rawExp :: DevicePtr a -> CSize -> DeviceM d ()

rawSqrt :: DevicePtr a -> CSize -> DeviceM d ()

rawSin :: DevicePtr a -> CSize -> DeviceM d ()

rawCos :: DevicePtr a -> CSize -> DeviceM d ()

rawTan :: DevicePtr a -> CSize -> DeviceM d ()

rawAsin :: DevicePtr a -> CSize -> DeviceM d ()

rawAcos :: DevicePtr a -> CSize -> DeviceM d ()

rawAtan :: DevicePtr a -> CSize -> DeviceM d ()

rawSinh :: DevicePtr a -> CSize -> DeviceM d ()

rawCosh :: DevicePtr a -> CSize -> DeviceM d ()

rawTanh :: DevicePtr a -> CSize -> DeviceM d ()

rawAsinh :: DevicePtr a -> CSize -> DeviceM d ()

rawAcosh :: DevicePtr a -> CSize -> DeviceM d ()

rawAtanh :: DevicePtr a -> CSize -> DeviceM d ()

rawPow :: DevicePtr a -> DevicePtr a -> CSize -> DeviceM d ()

rawMax :: DevicePtr a -> DevicePtr a -> CSize -> DeviceM d ()

broadcast_copy :: CInt -> CSize -> DevicePtr a -> DevicePtr CInt -> DevicePtr a -> DevicePtr CInt -> DeviceM d ()

generateUniform :: Generator d -> DevicePtr a -> CSize -> DeviceM d Status

Low-level CuRAND bindings.

generateNormal :: Generator d -> DevicePtr a -> CSize -> a -> a -> DeviceM d Status

generateLogNormal :: Generator d -> DevicePtr a -> CSize -> a -> a -> DeviceM d Status

data CFloat :: *

Haskell type representing the C float type.

data CDouble :: *

Haskell type representing the C double type.

mySizeOf :: forall a. Storable a => Proxy a -> Int