DeepBanana.Weights

Description

Weights for neural networks as heterogenous lists.

Documentation

newtype Weights a l

Weights for a neural network. Represents the weights as the concatenated HList of the individual layer's weights. The a parameter represents the weights scalar datatype, used for FixShape and VectorSpace instances.

Constructors

W

Fields

unWeights :: HList l

class Floating (FixScalar t) => FixShape t where

Class for datatypes which are not Floating nor VectorSpace instances themselves, but can be lifted to fixed-shape versions which are through an error monad. Somewhat ad-hoc class Used to implement stochastic gradient descent in Optimize for non fixed shape datatypes in a type safe manner.

Associated Types

type FixScalar t :: *

Methods

liftVec :: (MonadError e m, Variant e IncompatibleShape) => (forall t'. (Floating t', VectorSpace t', Scalar t' ~ FixScalar t) => t' -> t' -> t') -> t -> t -> m t

Re-exports

module DeepBanana.HList