DeepBanana.Tensor

Description

Immutable GPU multi-dimensional dense numeric arrays.

Documentation

data Tensor d s a

An immutable, dense numerical array stored on GPU device d, with shape s and scalar datatype a. As long as the shape datatype is fixed at the type level, it is an instance of all standard numeric type classes in an elementwise fashion.

Constructors

Tensor

Fields

shape :: s

dataptr :: !(ForeignPtr a)

Basic manipulations

dtype :: forall d s a. TensorScalar a => Tensor d s a -> DataType

Returns the CuDNN datatype of a tensor.

broadcast :: forall d t s1 s2 m a. (Device d, Shape s1, Shape s2, TensorScalar a, MonadError t m, Variant t IncompatibleShape, Variant t OutOfMemory) => s2 -> Tensor d s1 a -> m (Tensor d s2 a)

Type-safe broadcasting.

broadcast' :: forall s1 s2 d a. (Device d, Shape s1, Shape s2, TensorScalar a) => s2 -> Tensor d s1 a -> Tensor d s2 a

Unsafe version of broadcast, throws imprecise exceptions.

zeros :: (Device d, Shape s, TensorScalar a, MonadError t m, Variant t OutOfMemory) => s -> m (Tensor d s a)

Returns a tensor with all zero coefficients. Throws an OutOfMemory exception when the desired device is out of memory.

zeros' :: forall d s a. (Device d, Shape s, TensorScalar a) => s -> Tensor d s a

Returns a tensor with all zero coefficients. Throws an imprecise exception when the desired device is out of memory.

ones :: (Device d, Shape s, TensorScalar a, MonadError t m, Variant t OutOfMemory) => s -> m (Tensor d s a)

Initializes a tensor with all ones. Throws an OutOfMemory exception when the desired device is out of memory.

ones' :: forall d s a. (Device d, Shape s, TensorScalar a) => s -> Tensor d s a

Returns a tensor with all one coefficients. Throws an imprecise exception when the desired device is out of memory.

Shape manipulations

reshape :: (Shape s1, Shape s2, TensorScalar a, MonadError t m, Variant t IncompatibleSize) => s2 -> Tensor d s1 a -> m (Tensor d s2 a)

Reshaping. Throws an IncompatibleSize exception when the desired size and the input size do not correspond.

reshape' :: forall d s1 s2 a. (Shape s1, Shape s2, TensorScalar a) => s2 -> Tensor d s1 a -> Tensor d s2 a

Unsafe reshaping. Throws an imprecise exception when the desired size and the input size do not correspond.

shapeConvert :: forall s1 s2 d l a m t. (MonadError t m, Shape s1, Shape s2, Variant t IncompatibleShape) => s2 -> Tensor d s1 a -> m (Tensor d s2 a)

Shape conversion. This differs from reshape in that the desired shape and input shape must have exactly the same dimensions. This is useful when converting between dynamic and fixed shape datatypes. Throws an IncompatibleShape exception when the 2 shapes dimensions differ.

liftFixed1 :: forall d s1 a. Shape s1 => (forall s2. IsFixed s2 => Tensor d s2 a -> Tensor d s2 a) -> Tensor d s1 a -> Tensor d s1 a

Lifts a unary function on fixed shape tensors to work on any tensor. Useful to apply unary numeric functions such as log or exp to dynamic shape tensors in a type safe fashion.

liftFixed2 :: (Shape s1, MonadError t m, Variant t IncompatibleShape) => (forall s2. IsFixed s2 => Tensor d s2 a -> Tensor d s2 a -> Tensor d s2 a) -> Tensor d s1 a -> Tensor d s1 a -> m (Tensor d s1 a)

Lifts a binary function on fixed shape tensors to work on any tensor. Throws an IncompatibleShape exception when the input tensors do not have the same shape. Useful to apply binary numeric functions such as '(+)' and '(*)' to dynamic shape tensors in a type safe fashion.

Converting from/to mutable tensors

unsafeFreeze :: PrimMonad m => MTensor (PrimState m) d s a -> m (Tensor d s a)

Converts a mutable tensor into an immutable one in O(1) time. The data is not copied, so one should make sure the input mutable tensor is not modified after the the conversion. Unsafe.

unsafeThaw :: PrimMonad m => Tensor d s a -> m (MTensor (PrimState m) d s a)

Converts an immutable tensor into a mutable one in O(1) time. The data is not copied, so one should make sure the input immutable tensor is not used after the conversion. Unsafe.

Converting from/to lists

tensorFromList :: (Device d, Shape s, TensorScalar a, MonadError t m, Variant t IncompatibleSize, Variant t OutOfMemory) => s -> [a] -> m (Tensor d s a)

Initializes a tensor from list data. If the length of the list does not correspond to the size of the desired shape, throws an exception at run time.

tensorFromList' :: forall d s a. (Device d, Shape s, TensorScalar a) => s -> [a] -> Tensor d s a

Unsafe version of tensorFromList, throws imprecise exceptions.

tensorToList :: (Shape s, Device d, TensorScalar a) => Tensor d s a -> [a]

Converts a tensor to a list.

Converting from/to storable vectors

fromVector :: forall t m d s a. (Device d, Shape s, TensorScalar a, MonadError t m, Variant t OutOfMemory, Variant t IncompatibleSize) => s -> SVector a -> m (Tensor d s a)

Converts a storable vector to an immutable tensor. Throws an error at runtime when the input's length does not correspond to the desired output size. Runs in O(n) time.

fromVector' :: forall d s a. (Device d, Shape s, TensorScalar a) => s -> SVector a -> Tensor d s a

Unsafe version of fromVector, throws imprecise exceptions.

toVector :: forall d s a. (Device d, Shape s, TensorScalar a) => Tensor d s a -> SVector a

Converts a tensor to a storable vector. Runs in O(n) time.

Utilities

tconcat :: forall t d m a. (Device d, TensorScalar a, MonadError t m, Variant t OutOfMemory) => Tensor d (Dim 1) a -> Tensor d (Dim 1) a -> m (Tensor d (Dim 1) a)

Concatenates two 1-dimensional tensors. Inverse of tsplit. Runs in O(n+m) time.

tconcat' :: forall d a. (Device d, TensorScalar a) => Tensor d (Dim 1) a -> Tensor d (Dim 1) a -> Tensor d (Dim 1) a

Unsafe version of tconcat, throws imprecise exceptions.

tsplitAt :: forall t d m a. (Device d, TensorScalar a, MonadError t m, Variant t OutOfMemory, Variant t IncompatibleSize) => Int -> Tensor d (Dim 1) a -> m (Tensor d (Dim 1) a, Tensor d (Dim 1) a)

Splits a 1-dimensional tensor into 2 parts. Inverse of tconcat. Runs in O(n+m) time.

tsplitAt' :: forall d a. (Device d, TensorScalar a) => Int -> Tensor d (Dim 1) a -> (Tensor d (Dim 1) a, Tensor d (Dim 1) a)

Unsafe version of tSplit, throws imprecise exceptions.

elementwiseMax :: forall d s a t m. (Device d, IsFixed s, TensorScalar a, MonadError t m, Variant t OutOfMemory, Variant t IncompatibleShape) => Tensor d s a -> Tensor d s a -> m (Tensor d s a)

Computes the elementwise maximum of 2 tensors.

flatten :: forall s d a. (Shape s, TensorScalar a) => Tensor d s a -> Tensor d (Dim 1) a

Flattens a tensor into a 1-dimensional vector.

Re-exports

module DeepBanana.Tensor.Shape

module DeepBanana.Tensor.Exception

module DeepBanana.Tensor.TensorScalar