DeepBanana.Tensor.Shape

Description

Shape type class and instances. Defines the shape of a tensor, both at the type level and value level.

Shape type class

The Shape type class can only be implemented by types of kind '[Nat]'. These fully define the shape of a given tensor at the type level. Further, Shape also requires its instances to have size and number of dimension information available, both at the type level and the value level. These informations are used to ensure at compile time that most operations on tensor are correctly applied.

It also means that currently, we cannot handle tensors with (even partially) dynamic shapes. Although this is rarely an issue when training a neural network - which mostly expect inputs of a given size - 2d convolution layers are sometimes applied to inputs of varying width and height at test time. This should be adressed in the future.

class Shape s where

Type class for tensor shapes.

Minimal complete definition

shape

Associated Types

type Size s :: Nat

The size, i.e. the total number of elements of a tensor of shape s, at the type level. By convention, we set the empty type-level list to size 0.

type Nbdim s :: Nat

The number of dimensions of a tensor of shape s. For instance:

  Nbdim '[] ~ 0
  Nbdim '[5] ~ 1
  Nbdim '[6,8] ~ 2

Methods

shape :: Proxy s -> [Int]

Value-level shape of a tensor with type-level shape s.

nbdim :: Proxy s -> Int

Value-level number of dimensions of a tensor with type-level shape s.

size :: Proxy s -> Int

Value-level size, or number of elements of a tensor with type-level shape s. As with the type-level example, we set the empty type-level list to size 0.

Type-level arithmetic utilities

Some helper type families to compute mathematical operation in addition to the operations provided by GHC.TypeLits. Used for computing convolution and pooling output shape at compile time.

type family Quotient m n

Type level function to compute the quotient of the division of m by n.

  natVal (Proxy :: Proxy (Quotient m n)) == natVal (Proxy :: Proxy m) div natVal (Proxy :: Proxy n)

Equations

Quotient m n = Quotient' m n (CmpNat m n)

type family Remainder m n

Type level function to compute the remainder of the division of m by n.

  natVal (Proxy :: Proxy (Remainder m n)) == natVal (Proxy :: Proxy m) rem natVal (Proxy :: Proxy n)

Equations

Remainder m n = Remainder' m n (CmpNat m n)

type family Max m n

Type level max function for type-level natural numbers.

  natVal (Proxy :: Proxy (Max m n)) == max (natVal (Proxy :: Proxy m)) (natVal (Proxy :: Proxy n))

Equations

Max m n = Max' m n (CmpNat m n)